Extract DDL and cutover lifecycle tests (#716 subtasks 5+6)
Move 6 DDL algorithm selection tests into ddl_test.go:
- TestForNonInstantBurn
- TestIndexVisibility
- TestStatementWorkflowStillInstant
- TestTrailingSemicolon
- TestAlterExtendVarcharE2E
- TestMigrationWithSQLCommentsInStatement

Move 7 cutover lifecycle tests into cutover_test.go:
- TestSkipDropAfterCutoverLongTableName
- TestForRemainingTableArtifacts
- TestSkipDropAfterCutover
- TestDropAfterCutover
- TestDeferCutOver
- TestDeferCutOverE2E
- TestDeferCutOverE2EBinlogAdvance

Conversions:
- All table setup converted to NewTestTable
- All NewRunner calls converted to NewTestRunner/NewTestRunnerFromStatement/NewTestMigration
- assert converted to require (fail-fast)
- mysql.ParseDSN eliminated from all converted tests
- Cutover tests use WithSkipDropAfterCutover/WithDeferCutOver/WithRespectSentinel/WithDBName

runner_test.go: 1052 -> 323 lines (-69%)
ddl_test.go: 210 lines (new)
cutover_test.go: 418 -> 656 lines (+238, appended lifecycle tests)

Fixes #722, fixes #723
Part of #716 (test reorganization)
